Move IPMI handling into networking crate
