+++ rs/ic_os/config/Cargo.toml
@@ -14,6 +14,7 @@ serde = { workspace = true }
 serde_with = "1.6.2"
 regex = { workspace = true }
 mac_address = { path = "../network/mac_address" }
+network = { path = "../../network" } # Only required by bin
 
 [dev-dependencies]
 once_cell = "1.8"
