+++ rs/ic_os/config/src/main.rs
@@ -3,7 +3,8 @@ use clap::{Args, Parser, Subcommand};
 use config::config_ini::{get_config_ini_settings, ConfigIniSettings};
 use config::deployment_json::get_deployment_settings;
 use config::serialize_and_write_config;
-use mac_address::mac_address::{get_ipmi_mac, FormattedMacAddress};
+use mac_address::mac_address::FormattedMacAddress;
+use network::get_ipmi_mac;
 use regex::Regex;
 use std::fs::File;
 use std::path::{Path, PathBuf};
