+++ rs/ic_os/network/BUILD.bazel
@@ -10,6 +10,7 @@ DEPENDENCIES = [
     "@crate_index//:anyhow",
     "@crate_index//:ping",
     "@crate_index//:rayon",
+    "@crate_index//:regex",
     "@crate_index//:serde",
     "@crate_index//:serde_json",
     "@crate_index//:serde_with",
