+++ rs/ic_os/network/Cargo.toml
@@ -9,6 +9,7 @@ config = { path = "../config" }
 mac_address = { path = "./mac_address" }
 ping = { version = "^0.5.0" }
 rayon = { workspace = true }
+regex = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serde_with = { version = "^1.6.2" }
